🛡️ Sentinel: Harden ID generation with Edge-safe utility
- Created `src/lib/id-generator.ts` as a standalone, Edge-safe utility.
- Centralized ID generation for sessions, requests, correlation, and UI.
- Replaced insecure `Math.random()` and scattered `crypto.randomUUID()` calls.
- Standardized request ID generation across infrastructure modules.
- Ensured cross-runtime (Node.js/Edge/Browser) compatibility.
